+++ oe_patterns_cdk_common/efs.py
@@ -63,6 +63,13 @@ class Efs(Construct):
             self,
             "EfsSg",
             group_description="{}/Efs".format(Aws.STACK_NAME),
+            security_group_egress=[
+                aws_ec2.CfnSecurityGroup.EgressProperty(
+                    ip_protocol="-1",
+                    cidr_ip="0.0.0.0/0",
+                    description="all IPv4 egress traffic allowed"
+                )
+            ],
             vpc_id=vpc.id()
         )
         self.sg.override_logical_id(f"{id}Sg")