+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * ALB: Requiring CIDR range for AlbIngressCidr parameter
+* EFS: Explicit SG egress rule
+* ASG: Explicit SG egress rule
 
 # 3.0.2